- Fixed the initialization script to update correctly root and updated it to /bin/bash. - Fixed the initialization script to to encompass the lockout prb of the socket after the shutdown. - Added the installation of bush to Dockerfile. - Changed the base image to alpine:3.4 for layer compatibility

+++ startup.sh
@@ -1,14 +1,26 @@
-#!/bin/sh
+#!/bin/bash
 
-if [ -d /app/mysql ]; then
+set -eo pipefail
+shopt -s nullglob
+
+if [ ! -d "/run/mysqld" ]; then
+  mkdir -p /run/mysqld
+  chown -R mysql:mysql /run/mysqld
+  chmod 777 /run/mysqld 
+else  
+  rm -f /run/mysqld/msqld.sock
+fi
+
+if [ -d /data/mysql ]; then
   echo "[i] MySQL directory already present, skipping creation"
 else
   echo "[i] MySQL data directory not found, creating initial DBs"
 
-  mysql_install_db --user=root > /dev/null
+  #mysql_install_db --user=root > /dev/null
+  mysql_install_db > /dev/null
 
-  if [ "$MYSQL_ROOT_PASSWORD" = "" ]; then
-    MYSQL_ROOT_PASSWORD=111111
+  if [ "${MYSQL_ROOT_PASSWORD}" = "" ]; then
+    MYSQL_ROOT_PASSWORD="111111"
     echo "[i] MySQL root Password: $MYSQL_ROOT_PASSWORD"
   fi
 
@@ -16,22 +28,19 @@ else
   MYSQL_USER=${MYSQL_USER:-""}
   MYSQL_PASSWORD=${MYSQL_PASSWORD:-""}
 
-  if [ ! -d "/run/mysqld" ]; then
-    mkdir -p /run/mysqld
-  fi
-
   tfile=`mktemp`
+                                                                                                                                        
   if [ ! -f "$tfile" ]; then
-      return 1
+    return 1
   fi
 
   cat << EOF > $tfile
 USE mysql;
-FLUSH PRIVILEGES;
+DELETE FROM mysql.user;
+CREATE USER 'root'@'%' IDENTIFIED BY '${MYSQL_ROOT_PASSWORD}';
 GRANT ALL PRIVILEGES ON *.* TO 'root'@'%' WITH GRANT OPTION;
-UPDATE user SET password=PASSWORD("$MYSQL_ROOT_PASSWORD") WHERE user='root';
-GRANT ALL PRIVILEGES ON *.* TO 'root'@'localhost' WITH GRANT OPTION;
-UPDATE user SET password=PASSWORD("") WHERE user='root' AND host='localhost';
+DROP DATABASE IF EXISTS test;
+FLUSH PRIVILEGES;
 EOF
 
   if [ "$MYSQL_DATABASE" != "" ]; then
@@ -40,13 +49,42 @@ EOF
 
     if [ "$MYSQL_USER" != "" ]; then
       echo "[i] Creating user: $MYSQL_USER with password $MYSQL_PASSWORD"
-      echo "GRANT ALL ON \`$MYSQL_DATABASE\`.* to '$MYSQL_USER'@'%' IDENTIFIED BY '$MYSQL_PASSWORD';" >> $tfile
+      echo "CREATE USER '$MYSQL_USER'@'%' IDENTIFIED BY '$MYSQL_PASSWORD';" >> $tfile
+      echo "GRANT ALL ON \`$MYSQL_DATABASE\`.* to '$MYSQL_USER'@'%';" >> $tfile
     fi
   fi
 
-  /usr/bin/mysqld --user=root --bootstrap --verbose=0 < $tfile
+  # /usr/bin/mysqld --user=root --bootstrap < $tfile
+
+  /usr/bin/mysqld --skip-networking &
+  pid="$!"
+
+  mysql=( mysql --protocol=socket -uroot -hlocalhost --socket=/run/mysqld/mysqld.sock )
+
+  for i in {30..0}; do
+    if echo 'SELECT 1' | "${mysql[@]}" &> /dev/null; then
+       break
+    fi
+
+    echo 'MySQL init process in progress...'
+    sleep 1
+  done
+
+  if [ "$i" = 0 ]; then
+     echo >&2 'MySQL init process failed.'
+     exit 1
+  fi
+
+  "${mysql[@]}" < $tfile
+
   rm -f $tfile
+
+  if ! kill -s TERM "$pid" || ! wait "$pid"; then
+     echo >&2 'MySQL init process failed.'
+     exit 1
+  fi
 fi
 
+#exec /usr/bin/mysqld --user=root --console
 
-exec /usr/bin/mysqld --user=root --console
+exec /usr/bin/mysqld
